+++ src/lib/linear/linked-list/index.ts
@@ -1,1 +1,2 @@
 export * from './singly/SinglyLinkedList'
+export * from './doubly/DoublyLinkedList'