+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+### 0.0.3
+* Add doubly linked list: [PR](https://github.com/thomasbui93/cobi-data/pull/51)
+* Change Stack and Queue to use doubly linked list
+
 ### 0.0.2
 * Add binary heap
 * Add binary tree search